+++ module-operator/apis/platform/v1alpha1/module_types.go
@@ -86,26 +86,6 @@ type ModuleCondition struct {
 	Type LifecycleCondition `json:"type"`
 }
 
-type ChartVersion struct {
-	Name              string `json:"name"`
-	DefaultVersion    string `json:"defaultVersion,omitempty"`
-	SupportedVersions string `json:"supportedVersions,omitempty"`
-}
-
-type HelmChartRepository struct {
-	Name              string `json:"name"`
-	URI               string `json:"uri"`
-	CredentialsSecret string `json:"credentialsSecret,omitempty"`
-}
-
-type ModuleSourceRef struct {
-}
-
-type ModuleSource struct {
-	ChartRepo *HelmChartRepository `json:"repo,omitempty"`
-	SourceRef *ModuleSourceRef     `json:"sourceRef,omitempty"`
-}
-
 // ModuleClassType Identifies the lifecycle class used to manage a subset of Module types
 type ModuleClassType string
 
@@ -118,28 +98,11 @@ const (
 	CCMModuleClass    ModuleClassType = "ccm"
 )
 
-type HelmChart struct {
-	Name    string `json:"name"`
-	Version string `json:"version,omitempty"`
-	Path    string `json:"path,omitempty"`
-}
-
-type HelmRelease struct {
-	Name       string              `json:"name"`
-	Namespace  string              `json:"namespace,omitempty"`
-	ChartInfo  HelmChart           `json:"chart,omitempty"`
-	Repository HelmChartRepository `json:"repo,omitempty"`
-	Overrides  []Overrides         `json:"overrides,omitempty"`
-}
-
 type LifecycleCondition string
 
 const (
 	ConditionArrayLimit = 5
 
-	CondAlreadyInstalled    LifecycleCondition = "AlreadyInstalled"
-	CondAlreadyUninstalled  LifecycleCondition = "AlreadyUninstalled"
-	CondAlreadyUpgraded     LifecycleCondition = "AlreadyUpgraded"
 	CondPreInstall          LifecycleCondition = "PreInstall"
 	CondInstallStarted      LifecycleCondition = "InstallStarted"
 	CondInstallComplete     LifecycleCondition = "InstallComplete"
@@ -149,10 +112,6 @@ const (
 	CondPreUpgrade          LifecycleCondition = "PreUpgrade"
 	CondUpgradeStarted      LifecycleCondition = "UpgradeStarted"
 	CondUpgradeComplete     LifecycleCondition = "UpgradeComplete"
-	CondReady               LifecycleCondition = "Ready"
-	CondReconciling         LifecycleCondition = "Reconciling"
-	CondReconcilingComplete LifecycleCondition = "ReconcileComplete"
-	CondFailed              LifecycleCondition = "Failed"
 )
 
 func init() {
